+++ CHANGELOG.md
@@ -11,17 +11,29 @@ versioning principles. Unstable releases do not.
 ### [Unreleased]
 
 Breaking changes:
-* None.
+* `net-util`:
+  * `IncomingRequest.fromNodeRequest()` is now an `async` method, and its final
+    argument is now a catch-all `options`. ("If a function has more than two
+    arguments, you haven't discovered all of them yet." --Unknown)
 
 Other notable changes:
 * `net-util`:
-  * Fixed `IncomingRequest.getHeaderOrNull()`, which was very broken.
+  * `IncomingRequest`:
+    * Added `body` constructor option.
+    * Made it start rejecting requests whose request method (e.g. `GET`) isn't
+      defined to take a request body but where the request actually does have a
+      body.
+    * Fixed `getHeaderOrNull()`, which had been broken for a while.
+    * `fromNodeRequest()` now reads the request body when present, and returns
+      it in the constructed instance.
   * Defined a base class, `BaseResponse` for the two concrete response classes.
   * Added a handful of static getters to `StatusResponse`.
   * Various other tweaks and fixes, motivated by a downstream project.
 * `webapp-builtins`:
   * Make `StaticFiles` and `SimpleResponse` only respond successfully to `GET`
     and `HEAD` requests.
+* `webapp-core` / config:
+  * Added `NetworkEndpoint` configuration `maxRequestBodySize`.
 
 ### v0.7.8 -- 2024-07-30 -- stable release